pyproject.toml
[tool.poetry]
name = "eco"
version = "0.1.0"
description = ""
authors = ["atabekdemurtaza <mychannel21war@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
thor-devkit = "^1.0.14"
django-rest-framework = "^0.1.0"
pillow = "^10.2.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-filter = "^24.1"
drf-yasg = "^1.21.7"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
